move contract related logic to his own hook
